Background Tiles
- Added offset to chunkMap
- Bound key B to switch between foreground and background in freecam
- Added backChunkMap to GameWorld
- Added intensityMultiplier uniform to world shader
- Fixed unnecessary initial class for sand
- Updated world loader and saver to save the background tiles too
- Reverted back to old version number in world fragment shader (the new
was unnecessary)

+++ src/fr/iutvalence/info/dut/m2107/toolbox/Maths.java
@@ -115,7 +115,7 @@ public class Maths {
 	 * @param orientation the orientation to apply
 	 * @return a transformation matrix using x and y coordinates
 	 */
-	public static Matrix4f createTransformationMatrix(int x, int y, TileOrientation orientation) {
+	public static Matrix4f createTransformationMatrix(float x, float y, TileOrientation orientation) {
 		Matrix4f matrix = new Matrix4f();
 		matrix.setIdentity();
 		switch (orientation) {
